+++ staging/BrocadeLoomSplit/index.ts
@@ -1,0 +1,1 @@
+export { BrocadeLoomSplit, type BrocadeLoomSplitProps } from './BrocadeLoomSplit'