+++ staging/ZephyrGaleHero/index.ts
@@ -1,0 +1,1 @@
+export { ZephyrGaleHero, type ZephyrGaleHeroProps } from './ZephyrGaleHero'